kvm: Wrap qemu-img into it's own util
The KVM Agent used to run Script.runSimpleBashScript for various qemu-img
methods, but the output was not always checked properly.

This patch wraps qemu-img into it's own util to make integration more clean.

The goal is also to have more reliable executing since we can better check
if qemu-img did it's job.

Tests are also included to test the code.